Refactored all itinerary parsing to ./itin
